+++ app/cmd/rpc/common_test.go
@@ -540,6 +540,8 @@ func createTestACL(kp keys.KeyPair) govTypes.ACL {
 		acl.SetOwner("application/MaxApplications", kp.GetAddress())
 		acl.SetOwner("gov/daoOwner", kp.GetAddress())
 		acl.SetOwner("gov/upgrade", kp.GetAddress())
+		acl.SetOwner("application/MaximumChains", kp.GetAddress())
+		acl.SetOwner("pos/MaximumChains", kp.GetAddress())
 		testACL = acl
 	}
 	return testACL